Fix circular import (move fetching of GetCapabilities from LayersFactory to utils.js)
